[kea] fix likely typo
qui’ê is most certainly a typo: the ’ indicates elision, which implies that the i would be suppressed. Also, there are no other occurrences of this string in the text.

Where it gets more complicated is that there are 6 occurrences of qui ê (no elision) and 3 of qu’ê, and it is unclear which form is correct here. Aligning with the previous paragraph which has the same pattern.